+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Improvements
 * Refactor ChiSquared pdf to avoid raising 0 to negative exponent which is deprecated
 * Refactor LogLogistic pdf and cdf to avoid raising 0 to negative exponent which is deprecated
+* Improved implementation of NoncentralT `median`
 
 ## v2.12.0 - 2025-10-17